add exception handler
