+++ extractor/src/test/java/net/shamansoft/cookbook/CookbookControllerAdditionalTest.java
@@ -1,21 +1,19 @@
 package net.shamansoft.cookbook;
 
 import jakarta.servlet.http.HttpServletRequest;
-import net.shamansoft.cookbook.dto.Request;
 import net.shamansoft.cookbook.dto.RecipeResponse;
+import net.shamansoft.cookbook.dto.Request;
 import net.shamansoft.cookbook.service.Compressor;
+import net.shamansoft.cookbook.service.DriveService;
 import net.shamansoft.cookbook.service.RawContentService;
 import net.shamansoft.cookbook.service.TokenService;
 import net.shamansoft.cookbook.service.Transformer;
-import net.shamansoft.cookbook.service.DriveService;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-import org.junit.jupiter.api.extension.ExtendWith;
 
 import javax.naming.AuthenticationException;
 import java.io.IOException;
@@ -24,7 +22,11 @@ import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class CookbookControllerAdditionalTest {
@@ -60,7 +62,7 @@ class CookbookControllerAdditionalTest {
         // Mock fetch from URL when HTML is missing
         when(rawContentService.fetch(URL)).thenReturn(RAW_HTML);
         when(transformer.transform(RAW_HTML)).thenReturn(TRANSFORMED);
-        
+
         // Mock token flow not used in this scenario
         Map<String, String> headers = Map.of("X-S-AUTH-TOKEN", AUTH_TOKEN);
         when(driveService.getOrCreateFolder(AUTH_TOKEN)).thenReturn("folder123");
@@ -70,13 +72,13 @@ class CookbookControllerAdditionalTest {
 
         // Call controller method
         RecipeResponse response = controller.createRecipe(request, null, false, headers);
-        
+
         // Verify response
         assertThat(response.title()).isEqualTo(TITLE);
         assertThat(response.url()).isEqualTo(URL);
         assertThat(response.driveFileId()).isEqualTo("file123");
         assertThat(response.driveFileUrl()).isEqualTo("http://drive.example.com/file123");
-        
+
         // Verify interactions
         verify(rawContentService, times(1)).fetch(URL);
         verify(compressor, never()).decompress(any());
@@ -89,7 +91,7 @@ class CookbookControllerAdditionalTest {
         when(tokenService.getAuthToken(any(HttpHeaders.class))).thenReturn(AUTH_TOKEN);
         // Mock transformation
         when(transformer.transform(RAW_HTML)).thenReturn(TRANSFORMED);
-        
+
         // Mock token processing
         Map<String, String> headers = Map.of("X-S-AUTH-TOKEN", AUTH_TOKEN);
         when(driveService.getOrCreateFolder(AUTH_TOKEN)).thenReturn("folder123");
@@ -99,12 +101,12 @@ class CookbookControllerAdditionalTest {
 
         // Call controller method with "none" compression
         RecipeResponse response = controller.createRecipe(request, "none", false, headers);
-        
+
         // Verify response
         assertThat(response.title()).isEqualTo(TITLE);
         assertThat(response.url()).isEqualTo(URL);
         assertThat(response.driveFileId()).isEqualTo("file123");
-        
+
         // Verify that decompression was never called
         verify(compressor, never()).decompress(any());
         verify(rawContentService, never()).fetch(any());
@@ -114,7 +116,7 @@ class CookbookControllerAdditionalTest {
     void createRecipePropagatesIOExceptionFromDecompression() throws IOException {
         // Set up request with compressed HTML
         Request request = new Request(COMPRESSED_HTML, TITLE, URL);
-        
+
         // Mock decompression to throw exception
         when(compressor.decompress(COMPRESSED_HTML)).thenThrow(new IOException("Decompression error"));
 
@@ -123,8 +125,8 @@ class CookbookControllerAdditionalTest {
 
         // Assert that exception is thrown when called
         assertThatThrownBy(() -> controller.createRecipe(request, null, false, headers))
-            .isInstanceOf(IOException.class)
-            .hasMessageContaining("Decompression error");
+                .isInstanceOf(IOException.class)
+                .hasMessageContaining("Decompression error");
 
         // Verify no URL fetch was attempted
         verify(rawContentService, never()).fetch(any());
@@ -134,38 +136,18 @@ class CookbookControllerAdditionalTest {
     void createRecipeThrowsUnauthorizedWhenTokenIsInvalid() throws AuthenticationException {
         // Set up request
         Request request = new Request(RAW_HTML, TITLE, URL);
-        
+
         // Mock token verification to fail
         Map<String, String> headers = Map.of("X-S-AUTH-TOKEN", AUTH_TOKEN);
         when(tokenService.getAuthToken(any(HttpHeaders.class))).thenThrow(new AuthenticationException("Invalid auth token"));
 
         // Assert that unauthorized exception is thrown
         assertThatThrownBy(() -> controller.createRecipe(request, "none", false, headers))
-            .isInstanceOf(AuthenticationException.class);
+                .isInstanceOf(AuthenticationException.class);
 
         // Verify token service was called but drive service was not
         verify(driveService, never()).getOrCreateFolder(any());
         verify(transformer, never()).transform(any());
-    }
-    
-    @Test
-    void handleIOExceptionReturnsProperResponse() {
-        // Mock HTTP request
-        when(httpServletRequest.getRequestURI()).thenReturn("/recipe");
-        
-        // Create test exception
-        IOException testException = new IOException("Test error message");
-        
-        // Call exception handler
-        var response = controller.handleIOException(testException, httpServletRequest);
-        
-        // Verify response status and content
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-        assertThat(response.getBody()).isNotNull();
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("status", 400);
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("error", "IO Error");
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("message", "Test error message");
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("path", "/recipe");
     }
 
     @Test
@@ -174,85 +156,66 @@ class CookbookControllerAdditionalTest {
         Method extractHtmlMethod = CookbookController.class.getDeclaredMethod(
                 "extractHtml", Request.class, String.class);
         extractHtmlMethod.setAccessible(true);
-        
+
         // Set up request with compressed HTML
         Request request = new Request(COMPRESSED_HTML, TITLE, URL);
-        
+
         // Mock decompression to return raw HTML
         when(compressor.decompress(COMPRESSED_HTML)).thenReturn(RAW_HTML);
-        
+
         // Call private method via reflection
         String html = (String) extractHtmlMethod.invoke(controller, request, null);
-        
+
         // Verify result
         assertThat(html).isEqualTo(RAW_HTML);
-        
+
         // Verify compressor was called with correct argument
         verify(compressor).decompress(COMPRESSED_HTML);
         verify(rawContentService, never()).fetch(any());
     }
-    
+
     @Test
     void testExtractHtmlWithNoneCompression() throws Exception {
         // Get private method using reflection
         Method extractHtmlMethod = CookbookController.class.getDeclaredMethod(
                 "extractHtml", Request.class, String.class);
         extractHtmlMethod.setAccessible(true);
-        
+
         // Set up request with HTML
         Request request = new Request(RAW_HTML, TITLE, URL);
-        
+
         // Call private method via reflection with "none" compression
         String html = (String) extractHtmlMethod.invoke(controller, request, "none");
-        
+
         // Verify result
         assertThat(html).isEqualTo(RAW_HTML);
-        
+
         // Verify compressor was never called
         verify(compressor, never()).decompress(any());
         verify(rawContentService, never()).fetch(any());
     }
-    
+
     @Test
     void testExtractHtmlFallsBackToUrl() throws Exception {
         // Get private method using reflection
         Method extractHtmlMethod = CookbookController.class.getDeclaredMethod(
                 "extractHtml", Request.class, String.class);
         extractHtmlMethod.setAccessible(true);
-        
+
         // Set up request with only URL, no HTML
         Request request = new Request(null, TITLE, URL);
-        
+
         // Mock fetch from URL
         when(rawContentService.fetch(URL)).thenReturn(RAW_HTML);
-        
+
         // Call private method via reflection
         String html = (String) extractHtmlMethod.invoke(controller, request, null);
-        
+
         // Verify result
         assertThat(html).isEqualTo(RAW_HTML);
-        
+
         // Verify rawContentService was called with correct URL
         verify(rawContentService).fetch(URL);
         verify(compressor, never()).decompress(any());
-    }
-    
-    @Test
-    void handleGeneralExceptionReturnsInternalServerError() {
-        // Mock HTTP request
-        when(httpServletRequest.getRequestURI()).thenReturn("/recipe");
-        
-        // Create test exception
-        RuntimeException testException = new RuntimeException("Unexpected error");
-        
-        // Call exception handler
-        var response = controller.handleGeneralException(testException, httpServletRequest);
-        
-        // Verify response status and content
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
-        assertThat(response.getBody()).isNotNull();
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("status", 500);
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("error", "Internal Server Error");
-        assertThat(response.getBody()).hasFieldOrPropertyWithValue("path", "/recipe");
     }
 }
