+++ extractor/src/test/java/net/shamansoft/cookbook/CookbookControllerTest.java
@@ -1,7 +1,5 @@
 package net.shamansoft.cookbook;
 
-import jakarta.servlet.http.HttpServletRequest;
-import net.shamansoft.cookbook.dto.ErrorResponse;
 import net.shamansoft.cookbook.dto.RecipeResponse;
 import net.shamansoft.cookbook.dto.Request;
 import net.shamansoft.cookbook.service.Compressor;
@@ -9,20 +7,12 @@ import net.shamansoft.cookbook.service.DriveService;
 import net.shamansoft.cookbook.service.RawContentService;
 import net.shamansoft.cookbook.service.TokenService;
 import net.shamansoft.cookbook.service.Transformer;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.util.MultiValueMap;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.server.ResponseStatusException;
 
 import javax.naming.AuthenticationException;
 import java.io.IOException;
@@ -31,7 +21,6 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -54,7 +43,7 @@ class CookbookControllerTest {
 
     @InjectMocks
     private CookbookController controller;
-    
+
     private static final String TITLE = "Recipe Title";
     private static final String URL = "http://example.com";
     private static final String TOKEN = "token";
@@ -82,13 +71,13 @@ class CookbookControllerTest {
         when(driveService.getOrCreateFolder("token")).thenReturn("folder-id");
         when(driveService.generateFileName("Title")).thenReturn("file-name");
         when(driveService.uploadRecipeYaml("token", "folder-id", "file-name", "transformed content"))
-            .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
+                .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
 
         RecipeResponse response = controller.createRecipe(request, null, true, Map.of("X-S-AUTH-TOKEN", "token"));
 
         assertThat(response)
-            .extracting(RecipeResponse::title, RecipeResponse::url)
-            .containsExactly("Title", "http://example.com");
+                .extracting(RecipeResponse::title, RecipeResponse::url)
+                .containsExactly("Title", "http://example.com");
     }
 
     @Test
@@ -100,13 +89,13 @@ class CookbookControllerTest {
         when(driveService.getOrCreateFolder("token")).thenReturn("folder-id");
         when(driveService.generateFileName("Title")).thenReturn("file-name");
         when(driveService.uploadRecipeYaml("token", "folder-id", "file-name", "transformed content"))
-            .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
+                .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
 
         RecipeResponse response = controller.createRecipe(request, null, false, Map.of("X-S-AUTH-TOKEN", "token"));
 
         assertThat(response)
-            .extracting(RecipeResponse::title, RecipeResponse::url)
-            .containsExactly("Title", "http://example.com");
+                .extracting(RecipeResponse::title, RecipeResponse::url)
+                .containsExactly("Title", "http://example.com");
     }
 
     @Test
@@ -118,8 +107,8 @@ class CookbookControllerTest {
         when(driveService.getOrCreateFolder("token")).thenReturn("folder-id");
         when(driveService.generateFileName("Title")).thenReturn("file-name");
         when(driveService.uploadRecipeYaml("token", "folder-id", "file-name", "transformed content"))
-            .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
-        
+                .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
+
         RecipeResponse response = controller.createRecipe(request, null, true, Map.of("X-S-AUTH-TOKEN", "token"));
 
         assertThat(response.title()).isEqualTo("Title");
@@ -135,7 +124,7 @@ class CookbookControllerTest {
         when(driveService.getOrCreateFolder("token")).thenReturn("folder-id");
         when(driveService.generateFileName("Title")).thenReturn("file-name");
         when(driveService.uploadRecipeYaml("token", "folder-id", "file-name", "transformed content"))
-            .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
+                .thenReturn(new DriveService.UploadResult("file-id", "drive-url"));
 
         RecipeResponse response = controller.createRecipe(request, "none", false, Map.of("X-S-AUTH-TOKEN", "token"));
 
@@ -144,84 +133,34 @@ class CookbookControllerTest {
         assertThat(response.url()).isEqualTo("http://example.com");
     }
 
-    @Test
-    void ioExceptionFromDecompressionReturnsBadRequest() throws IOException {
-        HttpServletRequest mockRequest = mock(HttpServletRequest.class);
-        when(mockRequest.getRequestURI()).thenReturn("/recipe");
-        
-        ResponseEntity<Object> response = controller.handleIOException(
-            new IOException("Decompression failed"), mockRequest);
 
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-        
-        Object body = response.getBody();
-        assertThat(body).isInstanceOf(ErrorResponse.class);
-        
-        ErrorResponse errorResponse = (ErrorResponse) body;
-        assertThat(errorResponse.getMessage()).isEqualTo("Decompression failed");
-        assertThat(errorResponse.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(errorResponse.getError()).isEqualTo("IO Error");
-        assertThat(errorResponse.getPath()).isEqualTo("/recipe");
-    }
-
-    @Test
-    void validationExceptionReturnsBadRequest() {
-        MethodArgumentNotValidException ex = mock(MethodArgumentNotValidException.class);
-        BindingResult bindingResult = mock(BindingResult.class);
-        FieldError fieldError = mock(FieldError.class);
-        when(ex.getBindingResult()).thenReturn(bindingResult);
-        when(bindingResult.getFieldErrors()).thenReturn(java.util.Collections.singletonList(fieldError));
-        when(fieldError.getDefaultMessage()).thenReturn("Title is required");
-        when(fieldError.getField()).thenReturn("title");
-        
-        HttpServletRequest mockRequest = mock(HttpServletRequest.class);
-        when(mockRequest.getRequestURI()).thenReturn("/recipe");
-
-        ResponseEntity<Object> response = controller.handleValidationException(ex, mockRequest);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-        
-        Object body = response.getBody();
-        assertThat(body).isInstanceOf(ErrorResponse.class);
-        
-        ErrorResponse errorResponse = (ErrorResponse) body;
-        assertThat(errorResponse.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(errorResponse.getError()).isEqualTo("Validation Error");
-        assertThat(errorResponse.getPath()).isEqualTo("/recipe");
-        
-        assertThat(errorResponse.getValidationErrors()).hasSize(1);
-        ErrorResponse.ValidationError validationError = errorResponse.getValidationErrors().get(0);
-        assertThat(validationError.getField()).isEqualTo("title");
-        assertThat(validationError.getMessage()).isEqualTo("Title is required");
-    }
-    
     @Test
     void comprehensiveRecipeCreationTest() throws IOException, AuthenticationException {
         when(tokenService.getAuthToken(any(HttpHeaders.class))).thenReturn(TOKEN);
         // Setup request with all fields
         Request request = new Request("compressed content", TITLE, URL);
-        
+
         // Setup mocks for all services
         when(compressor.decompress("compressed content")).thenReturn(RAW_HTML);
         when(transformer.transform(RAW_HTML)).thenReturn(TRANSFORMED_CONTENT);
         when(driveService.getOrCreateFolder(TOKEN)).thenReturn(FOLDER_ID);
         when(driveService.generateFileName(TITLE)).thenReturn(FILE_NAME);
         when(driveService.uploadRecipeYaml(TOKEN, FOLDER_ID, FILE_NAME, TRANSFORMED_CONTENT))
-            .thenReturn(new DriveService.UploadResult("file-id-123", "https://drive.google.com/file-id-123"));
-        
+                .thenReturn(new DriveService.UploadResult("file-id-123", "https://drive.google.com/file-id-123"));
+
         // Create headers with auth token
         Map<String, String> headers = Map.of("X-S-AUTH-TOKEN", TOKEN);
-        
+
         // Execute controller method
         RecipeResponse response = controller.createRecipe(request, null, false, headers);
-        
+
         // Verify all service interactions
         verify(compressor).decompress("compressed content");
         verify(transformer).transform(RAW_HTML);
         verify(driveService).getOrCreateFolder(TOKEN);
         verify(driveService).generateFileName(TITLE);
         verify(driveService).uploadRecipeYaml(TOKEN, FOLDER_ID, FILE_NAME, TRANSFORMED_CONTENT);
-        
+
         // Assert response properties
         assertThat(response).isNotNull();
         assertThat(response.title()).isEqualTo(TITLE);
@@ -229,18 +168,18 @@ class CookbookControllerTest {
         assertThat(response.driveFileId()).isEqualTo("file-id-123");
         assertThat(response.driveFileUrl()).isEqualTo("https://drive.google.com/file-id-123");
     }
-    
+
     @Test
     void htmlDecompressionFailureWithNoUrlFallback() throws IOException {
         // Setup request with HTML content but no URL for fallback
         Request request = new Request("invalid-compressed-content", TITLE, null);
-        
+
         // Mock decompression to fail
         when(compressor.decompress("invalid-compressed-content")).thenThrow(new IOException("Invalid compressed data"));
-        
+
         // Create headers with auth token
         Map<String, String> headers = Map.of("X-S-AUTH-TOKEN", TOKEN);
-        
+
         // Assert that the correct exception is thrown
         assertThatThrownBy(() -> {
             try {
@@ -249,38 +188,13 @@ class CookbookControllerTest {
                 throw e;
             }
         })
-            .isInstanceOf(IOException.class)
-            .hasMessageContaining("Failed to decompress HTML and no valid URL provided as fallback");
-            
+                .isInstanceOf(IOException.class)
+                .hasMessageContaining("Failed to decompress HTML and no valid URL provided as fallback");
+
         // Verify decompression was attempted but HTML fetch was not
         verify(compressor).decompress("invalid-compressed-content");
         verify(rawContentService, never()).fetch(any());
     }
-    
-    @Test
-    void generalExceptionHandlerReturnsInternalServerError() {
-        // Setup mock request
-        HttpServletRequest mockRequest = mock(HttpServletRequest.class);
-        when(mockRequest.getRequestURI()).thenReturn("/recipe");
-        
-        // Create a runtime exception
-        RuntimeException exception = new RuntimeException("Something unexpected happened");
-        
-        // Call the exception handler
-        ResponseEntity<Object> response = controller.handleGeneralException(exception, mockRequest);
-        
-        // Verify response
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
-        
-        Object body = response.getBody();
-        assertThat(body).isInstanceOf(ErrorResponse.class);
-        
-        ErrorResponse errorResponse = (ErrorResponse) body;
-        assertThat(errorResponse.getStatus()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        assertThat(errorResponse.getError()).isEqualTo("Internal Server Error");
-        assertThat(errorResponse.getMessage()).isEqualTo("An unexpected error occurred. Please try again later.");
-        assertThat(errorResponse.getPath()).isEqualTo("/recipe");
-        // Error message should not expose internal details
-        assertThat(errorResponse.getMessage()).doesNotContain("Something unexpected happened");
-    }
+
+
 }
